+++ packages/core/src/utils/getStyle/index.ts
@@ -1,6 +1,14 @@
-import { ShallotProp } from '@shallot-ui/theme'
 import { CSSObject, DefaultTheme } from 'styled-components'
+import { ShallotProp } from '@shallot-ui/theme'
 
+/**
+ * Returns a function that generates CSS styles based on a given `shallot` object.
+ * The `shallot` object is a set of key-value pairs where the key is a CSS property and the value is either a string or a function that returns a string.
+ * If the value is a function, it will be called with the `theme` object as its argument.
+ * If the value is an object, `getStyle` will be called recursively on that object.
+ * @param shallot - An object containing CSS properties and values.
+ * @returns A function that generates CSS styles based on the `shallot` object.
+ */
 export const getStyle =
   <T extends { shallot?: ShallotProp }>({ shallot = {} }: T) =>
   ({ theme }: { theme: DefaultTheme }): CSSObject =>
@@ -21,6 +29,44 @@ export const getStyle =
       return acc
     }, {} as CSSObject)
 
+/**
+ * Defines the type of the state object for a Shallot component.
+ */
+type ShallotState = Record<string, ShallotProp>
+
+/**
+ * Defines the type of a Shallot prop that can contain a state object.
+ */
+export type StatefulShallotProp = Record<
+  string, // componentName
+  ShallotProp | ShallotState
+>
+
+/**
+ * Returns a function that generates CSS styles based on a nested `shallot` object.
+ * The `shallot` object is a set of key-value pairs where the key is a CSS property and the value is either a string or a function that returns a string.
+ * If the value is a function, it will be called with the `theme` object as its argument.
+ * If the value is an object, `getStyle` will be called recursively on that object.
+ * @param componentName - The name of the component to generate styles for.
+ * @param state - The name of the state object to generate styles for.
+ * @returns A function that generates CSS styles based on the nested `shallot` object.
+ */
+export const getNestedStyle =
+  (componentName: string, state?: string) =>
+  <T extends { shallot?: StatefulShallotProp }>({ shallot = {} }: T) =>
+  ({ theme }: { theme: DefaultTheme }): CSSObject => {
+    const nestedShallot = (state ? shallot?.[state] : shallot)?.[componentName]
+    return getStyle({ shallot: nestedShallot as ShallotProp })({ theme })
+  }
+
+/**
+ * Returns a function that generates CSS styles based on a `shallot` object with optional variant styles.
+ * The `shallot` object is a set of key-value pairs where the key is a CSS property and the value is either a string or a function that returns a string.
+ * If the value is a function, it will be called with the `theme` object as its argument.
+ * If the value is an object, `getStyle` will be called recursively on that object.
+ * @param variantNamespace - The namespace of the variant styles to use.
+ * @returns A function that generates CSS styles based on the `shallot` object with optional variant styles.
+ */
 export const scopeGetStyle =
   (variantNamespace?: string) =>
   <T extends { shallot?: {}; variant?: string }>({