+++ platforms/native/components/ui/index.ts
@@ -1,0 +1,4 @@
+export * from './Image'
+export * from './Button'
+export * from './Input'
+export * from './Tag'